moved ChatRoom into types
