--- src/lib/ChatRoom.ts
@@ -1,16 +1,0 @@
-import { Stream } from 'streamr-client'
-
-export interface ChatMessage {
-    type: 'text' | 'metadata'
-    payload: string
-}
-
-export interface ChatRoom {
-    id: string
-    name: string
-    stream: Stream
-    publishMessage: (message: string) => Promise<void>
-    publishMetadata: (metadata: any) => Promise<void>
-    subscribeMessages: (callback: (message: ChatMessage) => void) => void
-    subscribeMetadata: (callback: (metadata: any) => void) => void
-}
